Module dual_iir::app

The RTIC application module

Modules

eth

Hardware task

ethernet_link

Software task

idle

Idle loop

init

Initialization function

process

Hardware task

settings_update

Software task

shared_resources 🔒

spi2

Hardware task

spi3

Hardware task

spi4

Hardware task

spi5

Hardware task

start

Software task

telemetry

Software task

usb

Software task

Structs

Local 🔒

Local resources

Shared 🔒

User code end Shared resources

__rtic_internal_eth_Context

Execution context

__rtic_internal_ethernet_linkSharedResources

Shared resources ethernet_link has access to

__rtic_internal_ethernet_link_Context

Execution context

__rtic_internal_idleSharedResources

Shared resources idle has access to

__rtic_internal_idle_Context

Execution context

__rtic_internal_init_Context

Execution context

__rtic_internal_processLocalResources

Local resources process has access to

__rtic_internal_processSharedResources

Shared resources process has access to

__rtic_internal_process_Context

Execution context

__rtic_internal_settings_updateLocalResources

Local resources settings_update has access to

__rtic_internal_settings_updateSharedResources

Shared resources settings_update has access to

__rtic_internal_settings_update_Context

Execution context

__rtic_internal_spi2_Context

Execution context

__rtic_internal_spi3_Context

Execution context

__rtic_internal_spi4_Context

Execution context

__rtic_internal_spi5_Context

Execution context

__rtic_internal_startLocalResources

Local resources start has access to

__rtic_internal_start_Context

Execution context

__rtic_internal_telemetryLocalResources

Local resources telemetry has access to

__rtic_internal_telemetrySharedResources

Shared resources telemetry has access to

__rtic_internal_telemetry_Context

Execution context

__rtic_internal_usbLocalResources

Local resources usb has access to

__rtic_internal_usbSharedResources

Shared resources usb has access to

__rtic_internal_usb_Context

Execution context

Statics

RTIC_ASYNC_MAX_LOGICAL_PRIO 🔒

Holds the maximum priority level for use by async HAL drivers.

__rtic_internal_ethernet_link_EXEC 🔒

__rtic_internal_settings_update_EXEC 🔒

__rtic_internal_start_EXEC 🔒

__rtic_internal_telemetry_EXEC 🔒

__rtic_internal_usb_EXEC 🔒

Functions

DMA1_STR4 🔒 ⚠

ETH 🔒 ⚠

SDMMC 🔒 ⚠

Interrupt handler to dispatch async tasks at priority 1

SPI2 🔒 ⚠

SPI3 🔒 ⚠

SPI4 🔒 ⚠

SPI5 🔒 ⚠

eth 🔒

ethernet_link 🔒

idle 🔒

init 🔒

process 🔒

settings_update 🔒

spi2 🔒

spi3 🔒

spi4 🔒

spi5 🔒

start 🔒

telemetry 🔒

usb 🔒