Crate dual_iir

Dual IIR

The Dual IIR application exposes two configurable channels. Stabilizer samples input at a fixed rate, digitally filters the data, and then generates filtered output signals on the respective channel outputs.

Features

Two indpenendent channels

up to 800 kHz rate, timed sampling

Run-time filter configuration

Input/Output data streaming

Down to 2 µs latency

f32 IIR math

Generic biquad (second order) IIR filter

Anti-windup

Derivative kick avoidance

Settings

Refer to the DualIir structure for documentation of run-time configurable settings for this application.

Telemetry

Refer to Telemetry for information about telemetry reported by this application.

Livestreaming

This application streams raw ADC and DAC data over UDP. Refer to stabilizer::net::data_stream for more information.

Modules

app

The RTIC application module

Structs

DualIir

Settings

Constants

BATCH_SIZE 🔒

IIR_CASCADE_LENGTH 🔒

SAMPLE_PERIOD 🔒

SAMPLE_TICKS 🔒

SAMPLE_TICKS_LOG2 🔒

SCALE 🔒