Module dual_iir::app

The RTIC application module

Re-exports

pub use rtic::Monotonic as _;

Modules

eth

Hardware task

ethernet_link

Software task

idle

Idle loop

init

Initialization function

monotonics

Holds static methods for each monotonic.

process

Hardware task

settings_update

Software task

shared_resources

spi2

Hardware task

spi3

Hardware task

spi4

Hardware task

spi5

Hardware task

telemetry_task

Software task

Structs

Local

Shared

__rtic_internal_Monotonics

Monotonics used by the system

__rtic_internal_eth_Context

Execution context

__rtic_internal_ethernet_linkSharedResources

Shared resources ethernet_link has access to

__rtic_internal_ethernet_link_Context

Execution context

__rtic_internal_ethernet_link_Monotonic_SpawnHandle

__rtic_internal_idleSharedResources

Shared resources idle has access to

__rtic_internal_idle_Context

Execution context

__rtic_internal_init_Context

Execution context

__rtic_internal_processLocalResources

Local resources process has access to

__rtic_internal_processSharedResources

Shared resources process has access to

__rtic_internal_process_Context

Execution context

__rtic_internal_settings_updateLocalResources

Local resources settings_update has access to

__rtic_internal_settings_updateSharedResources

Shared resources settings_update has access to

__rtic_internal_settings_update_Context

Execution context

__rtic_internal_settings_update_Monotonic_SpawnHandle

__rtic_internal_spi2_Context

Execution context

__rtic_internal_spi3_Context

Execution context

__rtic_internal_spi4_Context

Execution context

__rtic_internal_spi5_Context

Execution context

__rtic_internal_telemetry_taskSharedResources

Shared resources telemetry_task has access to

__rtic_internal_telemetry_task_Context

Execution context

__rtic_internal_telemetry_task_Monotonic_SpawnHandle

Functions

DCMI⚠

Interrupt handler to dispatch tasks at priority 1

DMA1_STR4⚠

Main DSP processing routine for Stabilizer.

ETH⚠

SPI2⚠

SPI3⚠

SPI4⚠

SPI5⚠

TIM15⚠

__rtic_internal_ethernet_link_Monotonic_spawn_after

Spawns the task after a set duration relative to the current time

__rtic_internal_ethernet_link_Monotonic_spawn_at

Spawns the task at a fixed time instant

__rtic_internal_ethernet_link_spawn

Spawns the task directly

__rtic_internal_settings_update_Monotonic_spawn_after

Spawns the task after a set duration relative to the current time

__rtic_internal_settings_update_Monotonic_spawn_at

Spawns the task at a fixed time instant

__rtic_internal_settings_update_spawn

Spawns the task directly

__rtic_internal_telemetry_task_Monotonic_spawn_after

Spawns the task after a set duration relative to the current time

__rtic_internal_telemetry_task_Monotonic_spawn_at

Spawns the task at a fixed time instant

__rtic_internal_telemetry_task_spawn

Spawns the task directly

eth

ethernet_link

idle

init

User code end

process

Main DSP processing routine for Stabilizer.

settings_update

spi2

spi3

spi4

spi5

telemetry_task

Type Definitions

Monotonic

User code from within the module