dual_iir
Struct DualIir

pub struct DualIir {
    pub(crate) afe: [Leaf<Gain>; 2],
    pub(crate) iir_ch: [[Leaf<Biquad<f32>>; 1]; 2],
    pub(crate) allow_hold: Leaf<bool>,
    pub(crate) force_hold: Leaf<bool>,
    pub(crate) telemetry_period: Leaf<u16>,
    pub(crate) stream: Leaf<StreamTarget>,
    pub(crate) source: [BasicConfig; 2],
}

Fields
afe: [Leaf<Gain>; 2]

Configure the Analog Front End (AFE) gain.

Path

afe/<n>

<n> specifies which channel to configure. <n> := [0, 1]

Value

Any of the variants of Gain enclosed in double quotes.

iir_ch: [[Leaf<Biquad<f32>>; 1]; 2]

Configure the IIR filter parameters.

Path

iir_ch/<n>/<m>

<n> specifies which channel to configure. <n> := [0, 1]

<m> specifies which cascade to configure. <m> := [0, 1], depending on IIR_CASCADE_LENGTH

See iir::Biquad

allow_hold: Leaf<bool>

Specified true if DI1 should be used as a “hold” input.

Path

allow_hold

Value

“true” or “false”

force_hold: Leaf<bool>

Specified true if “hold” should be forced regardless of DI1 state and hold allowance.

Path

force_hold

Value

“true” or “false”

telemetry_period: Leaf<u16>

Specifies the telemetry output period in seconds.

Path

telemetry_period

Value

Any non-zero value less than 65536.

stream: Leaf<StreamTarget>

Specifies the target for data streaming.

Path

stream

Value

See StreamTarget

source: [BasicConfig; 2]

Specifies the config for signal generators to add on to DAC0/DAC1 outputs.

Path

source/<n>

<n> specifies which channel to configure. <n> := [0, 1]

Value

See signal_generator::BasicConfig

Implementations

impl DualIir

pub(crate) const __MINICONF_LOOKUP: KeyLookup = _