Struct dual_iir::DualIir

pub struct DualIir {
    pub(crate) afe: [Gain; 2],
    pub(crate) iir_ch: [[Biquad<f32>; 1]; 2],
    pub(crate) allow_hold: bool,
    pub(crate) force_hold: bool,
    pub(crate) telemetry_period: u16,
    pub(crate) stream_target: StreamTarget,
    pub(crate) signal_generator: [BasicConfig; 2],
}

Fields
afe: [Gain; 2]

Configure the Analog Front End (AFE) gain.

Path

afe/<n>

<n> specifies which channel to configure. <n> := [0, 1]

Value

Any of the variants of Gain enclosed in double quotes.

iir_ch: [[Biquad<f32>; 1]; 2]

Configure the IIR filter parameters.

Path

iir_ch/<n>/<m>

<n> specifies which channel to configure. <n> := [0, 1]

<m> specifies which cascade to configure. <m> := [0, 1], depending on IIR_CASCADE_LENGTH

See iir::Biquad

allow_hold: bool

Specified true if DI1 should be used as a “hold” input.

Path

allow_hold

Value

“true” or “false”

force_hold: bool

Specified true if “hold” should be forced regardless of DI1 state and hold allowance.

Path

force_hold

Value

“true” or “false”

telemetry_period: u16

Specifies the telemetry output period in seconds.

Path

telemetry_period

Value

Any non-zero value less than 65536.

stream_target: StreamTarget

Specifies the target for data livestreaming.

Path

stream_target

Value

See StreamTarget

signal_generator: [BasicConfig; 2]

Specifies the config for signal generators to add on to DAC0/DAC1 outputs.

Path

signal_generator/<n>

<n> specifies which channel to configure. <n> := [0, 1]

Value

See signal_generator::BasicConfig

Implementations

impl DualIir

pub(crate) fn __miniconf_lookup<K: Keys>( keys: &mut K ) -> Result<usize, Traversal>